# *: record and print the plan in slow log. (#12179)