--- a/planner/core/common_plans.go
+++ b/planner/core/common_plans.go
@@ -34,6 +34,7 @@ import (
 	driver "github.com/pingcap/tidb/types/parser_driver"
 	"github.com/pingcap/tidb/util/chunk"
 	"github.com/pingcap/tidb/util/kvcache"
+	"github.com/pingcap/tidb/util/plancodec"
 	"github.com/pingcap/tidb/util/ranger"
 	"github.com/pingcap/tidb/util/texttree"
 )
@@ -645,6 +646,53 @@ func (e *Explain) prepareOperatorInfo(p PhysicalPlan, taskType string, indent st
 		}
 	}
 	e.Rows = append(e.Rows, row)
+}
+
+func (e *Explain) prettyIdentifier(id, indent string, isLastChild bool) string {
+	if len(indent) == 0 {
+		return id
+	}
+
+	indentBytes := []rune(indent)
+	for i := len(indentBytes) - 1; i >= 0; i-- {
+		if indentBytes[i] != plancodec.TreeBody {
+			continue
+		}
+
+		// Here we attach a new node to the current sub-tree by changing
+		// the closest TreeBody to a:
+		// 1. TreeLastNode, if this operator is the last child.
+		// 2. TreeMiddleNode, if this operator is not the last child..
+		if isLastChild {
+			indentBytes[i] = plancodec.TreeLastNode
+		} else {
+			indentBytes[i] = plancodec.TreeMiddleNode
+		}
+		break
+	}
+
+	// Replace the TreeGap between the TreeBody and the node to a
+	// TreeNodeIdentifier.
+	indentBytes[len(indentBytes)-1] = plancodec.TreeNodeIdentifier
+	return string(indentBytes) + id
+}
+
+func (e *Explain) getIndent4Child(indent string, isLastChild bool) string {
+	if !isLastChild {
+		return string(append([]rune(indent), plancodec.TreeBody, plancodec.TreeGap))
+	}
+
+	// If the current node is the last node of the current operator tree, we
+	// need to end this sub-tree by changing the closest TreeBody to a TreeGap.
+	indentBytes := []rune(indent)
+	for i := len(indentBytes) - 1; i >= 0; i-- {
+		if indentBytes[i] == plancodec.TreeBody {
+			indentBytes[i] = plancodec.TreeGap
+			break
+		}
+	}
+
+	return string(append(indentBytes, plancodec.TreeBody, plancodec.TreeGap))
 }
 
 func (e *Explain) prepareDotInfo(p PhysicalPlan) {