### What problem does this PR solve? <!--add issue link with summary if exists-->
Backport #12179 by cherry-picking #12808 

Conflict files(from #12808):
```
        both modified:   executor/adapter.go
        both modified:   planner/core/common_plans.go
        both modified:   sessionctx/variable/session.go
```

Related Parser PR: https://github.com/pingcap/parser/pull/629

---------------------------------------------
<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
1. Print the plan in slow log.
2. Add `slow-log-plan` config item and system variable to control enable.
3. Add plan encode and plan decoder function.
4. Add `tidb_decode_plan` function to decode the ecoded plan string.
5. Add `Plan` column to `Slow_query` table.

Related PR: https://github.com/pingcap/parser/pull/557

**Example**

Enable `tidb_record_plan_in_slow_log` first.

```sql
set @@tidb_record_plan_in_slow_log=1;
select count(*) from t1  join t1 t2 where t1.a=t2.b union select count(*) from t1  join t1 t2 where t1.a=t2.b union select count(*) from t1  join t1 t2 where t1.a=t2.b union select sleep(1);
```

**In slow log**

As is seen, `Plan` is the string of the encoded query plan.

```
# Time: 2019-09-12T19:36:50.003062+08:00
# Txn_start_ts: 411117344255639553
# User: root@127.0.0.1
# Conn_ID: 1
# Query_time: 1.002325096
# Request_count: 6 Total_keys: 36 Process_keys: 30
# DB: test
# Index_names: [t1:a,t1:b,t1:a,t1:b,t1:a,t1:b]
# Is_internal: false
# Digest: 35476665610eec1a5dff15ef68287edc48fa6b924d650f8dc26c4754b2293322
# Stats: t1:pseudo
# Num_cop_tasks: 6
# Cop_proc_avg: 0 Cop_proc_p90: 0 Cop_proc_max: 0 Cop_proc_addr: 127.0.0.1:20164
# Cop_wait_avg: 0 Cop_wait_p90: 0 Cop_wait_max: 0 Cop_wait_addr: 127.0.0.1:20164
# Mem_max: 2232
# Succ: true
# Plan: tidb_decode_plan('eNq0lM1uqzAQRtfDU3iZ3EsRY2MT+w267rKqLCAmpYkMso3SvH3FnxpVIouKbljMN8xhjkakIDSjkEIGJ9f2HSlvqmp7G3b/9jGpe1t5VTfOB9ded0uwjxCkZmx8LaLANMsgBYSq8OGuiwHXTI7JNGmKcB9lgFKLIRIJzVjOobHWOPLRNjYmF1MHcjY3FYwPScCkiIlrTu/3RZqUEQdGtRi/PpFymHE0n+p5eL5UhdUCIwHItEDAuSUU5cWogDGZmofJhT0Z9frU2Dr+39j6LSZnYzrSuqNxKrjexMSHInjVedMf2xmbrWPZjGU/sXTBlr/ADp5luuZZ8lXPSPkWohEPqysj5tPOiPn2rpGmD8hyIcvNdWMm1nwjx3XhB9xEeL5+Y5jPR4Y5+wPhuXhA5guZby6cpvRb+CB6KLGx5C/GdJNiyjVNpz+Oa69eYfQVAAD//z6BX5c=')
select count(*) from t1  join t1 t2 where t1.a=t2.b union select count(*) from t1  join t1 t2 where t1.a=t2.b union select count(*) from t1  join t1 t2 where t1.a=t2.b union select sleep(1);
```

**query the `slow_query` table**

```sql
mysql>select plan,query from slow_query where plan != "" order by query_time  limit 1;
+-------------------------------------------------------------------------------------------------------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| plan                                                                                                        | query                                                                                                                                                                                          |
+-------------------------------------------------------------------------------------------------------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
|       HashAgg_32                      root    4       group by:Column#32, funcs:firstrow(Column#32)                      | select count(*) from t1  join t1 t2 where t1.a=t2.b union select count(*) from t1  join t1 t2 where t1.a=t2.b union select count(*) from t1  join t1 t2 where t1.a=t2.b union select sleep(1); |
|       └─Union_33                      root    4                                                                          |                                                                                                                                                                                                |
|         ├─Projection_34               root    1       cast(Column#30)                                                    |                                                                                                                                                                                                |
|         │ └─HashAgg_37                root    1       funcs:count(1)                                                     |                                                                                                                                                                                                |
|         │   └─HashLeftJoin_40         root    12487.5 inner join, inner:TableReader_47, equal:[eq(Column#22, Column#27)] |                                                                                                                                                                                                |
|         │     ├─TableReader_44        root    9990    data:Selection_43                                                  |                                                                                                                                                                                                |
|         │     │ └─Selection_43        cop     9990    not(isnull(Column#22))                                             |                                                                                                                                                                                                |
|         │     │   └─TableScan_42      cop     10000   table:t1, range:[-inf,+inf], keep order:false, stats:pseudo        |                                                                                                                                                                                                |
|         │     └─TableReader_47        root    9990    data:Selection_46                                                  |                                                                                                                                                                                                |
|         │       └─Selection_46        cop     9990    not(isnull(Column#27))                                             |                                                                                                                                                                                                |
|         │         └─TableScan_45      cop     10000   table:t2, range:[-inf,+inf], keep order:false, stats:pseudo        |                                                                                                                                                                                                |
|         ├─Projection_50               root    1       cast(Column#20)                                                    |                                                                                                                                                                                                |
|         │ └─HashAgg_53                root    1       funcs:count(1)                                                     |                                                                                                                                                                                                |
|         │   └─HashLeftJoin_56         root    12487.5 inner join, inner:TableReader_63, equal:[eq(Column#12, Column#17)] |                                                                                                                                                                                                |
|         │     ├─TableReader_60        root    9990    data:Selection_59                                                  |                                                                                                                                                                                                |
|         │     │ └─Selection_59        cop     9990    not(isnull(Column#12))                                             |                                                                                                                                                                                                |
|         │     │   └─TableScan_58      cop     10000   table:t1, range:[-inf,+inf], keep order:false, stats:pseudo        |                                                                                                                                                                                                |
|         │     └─TableReader_63        root    9990    data:Selection_62                                                  |                                                                                                                                                                                                |
|         │       └─Selection_62        cop     9990    not(isnull(Column#17))                                             |                                                                                                                                                                                                |
|         │         └─TableScan_61      cop     10000   table:t2, range:[-inf,+inf], keep order:false, stats:pseudo        |                                                                                                                                                                                                |
|         ├─Projection_66               root    1       cast(Column#10)                                                    |                                                                                                                                                                                                |
|         │ └─HashAgg_69                root    1       funcs:count(1)                                                     |                                                                                                                                                                                                |
|         │   └─HashLeftJoin_72         root    12487.5 inner join, inner:TableReader_79, equal:[eq(Column#2, Column#7)]   |                                                                                                                                                                                                |
|         │     ├─TableReader_76        root    9990    data:Selection_75                                                  |                                                                                                                                                                                                |
|         │     │ └─Selection_75        cop     9990    not(isnull(Column#2))                                              |                                                                                                                                                                                                |
|         │     │   └─TableScan_74      cop     10000   table:t1, range:[-inf,+inf], keep order:false, stats:pseudo        |                                                                                                                                                                                                |
|         │     └─TableReader_79        root    9990    data:Selection_78                                                  |                                                                                                                                                                                                |
|         │       └─Selection_78        cop     9990    not(isnull(Column#7))                                              |                                                                                                                                                                                                |
|         │         └─TableScan_77      cop     10000   table:t2, range:[-inf,+inf], keep order:false, stats:pseudo        |                                                                                                                                                                                                |
|         └─Projection_82               root    1       cast(Column#1)                                                     |                                                                                                                                                                                                |
|           └─Projection_83             root    1       sleep(1)                                                           |                                                                                                                                                                                                |
|             └─TableDual_84            root    1       rows:1                                                             |                                                                                                                                                                                                |
+-------------------------------------------------------------------------------------------------------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
```

**Use `tidb_decode_plan` function to decode the encoded plan string.**

```sql
mysql>select tidb_decode_plan('jAu4MAk2XzMyCTAJNAlncm91cCBieTpDb2x1bW4jMzIsIGZ1bmNzOmZpcnN0cm93KEMRGhwpCjEJOV8zMwU5PAoyCTNfMzQJMAkxCWNhc3QVJwwwKQozAWAANwUbCUw8Y291bnQoMSkKNAkxN180MAEbSDI0ODcuNQlpbm5lciBqb2luLCAFDGQ6VGFibGVSZWFkZXJfNDcsIGVxdWFsOltlcRFjDDIyLCANrygyNyldCjUJMzJfNAGLeDk5OTAJZGF0YTpTZWxlY3Rpb25fNDMKNgkxXzQzCTEJICRub3QoaXNudWxsGVFMKSkKNwkxMF80MgkxCTEwMDAwCXQBh9g6dDEsIHJhbmdlOlstaW5mLCtpbmZdLCBrZWVwIG9yZGVyOmZhbHNlLCBzdGF0czpwc2V1ZG8KCZIhAlKSAAA2CZIANmqSAAA3FZIANT6SAAAyzpIAIa4ANSV4Mq4BADItrgA1IdUAMVKuAQQ1NqKuAQQ2M0quAQAxOa4AMTWuADYBi04cAQQ1OSUcCDU5CWKuAQAxNa4ENTg+HAHmrgEANiECTpIABDYyBZIANkEaBSBGQAIAMTWuBDYxPpIA4q4BADYleDKuAQAxLa4ENjkhk1KuAQA3YdeWXAMENzlKrgE5rTWsADcBiQX6OloDCDc1CmFaADdBogUgRhoBNasENzQ+GQHmqwEANwH/BXE+kQAAOAmRQRcFIEaRADWqBDc3PpEA4qoBADgldDaqAaEFCDNfOEFVGDEJc2xlZXCJ/0AyNl84NAkwCTEJcm93czoxCg==');
+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| tidb_decode_plan('jAu4MAk2XzMyCTAJNAlncm91cCBieTpDb2x1bW4jMzIsIGZ1bmNzOmZpcnN0cm93KEMRGhwpCjEJOV8zMwU5PAoyCTNfMzQJMAkxCWNhc3QVJwwwKQozAWAANwUbCUw8Y291bnQoMSkKNAkxN180MAEbSDI0ODcuNQlpbm5lciBqb2luLCAFDGQ6VGFibGVSZWFkZXJfNDcsIGVxdWFsOltlcRFjDDIyLCANrygyNyldCj |
+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
|       HashAgg_32                      root    4       group by:Column#32, funcs:firstrow(Column#32)                                                                                                                                                                           |
|       └─Union_33                      root    4                                                                                                                                                                                                                               |
|         ├─Projection_34               root    1       cast(Column#30)                                                                                                                                                                                                         |
|         │ └─HashAgg_37                root    1       funcs:count(1)                                                                                                                                                                                                          |
|         │   └─HashLeftJoin_40         root    12487.5 inner join, inner:TableReader_47, equal:[eq(Column#22, Column#27)]                                                                                                                                                      |
|         │     ├─TableReader_44        root    9990    data:Selection_43                                                                                                                                                                                                       |
|         │     │ └─Selection_43        cop     9990    not(isnull(Column#22))                                                                                                                                                                                                  |
|         │     │   └─TableScan_42      cop     10000   table:t1, range:[-inf,+inf], keep order:false, stats:pseudo                                                                                                                                                             |
|         │     └─TableReader_47        root    9990    data:Selection_46                                                                                                                                                                                                       |
|         │       └─Selection_46        cop     9990    not(isnull(Column#27))                                                                                                                                                                                                  |
|         │         └─TableScan_45      cop     10000   table:t2, range:[-inf,+inf], keep order:false, stats:pseudo                                                                                                                                                             |
|         ├─Projection_50               root    1       cast(Column#20)                                                                                                                                                                                                         |
|         │ └─HashAgg_53                root    1       funcs:count(1)                                                                                                                                                                                                          |
|         │   └─HashLeftJoin_56         root    12487.5 inner join, inner:TableReader_63, equal:[eq(Column#12, Column#17)]                                                                                                                                                      |
…